refactor: update from PR feedback
