+++ packages/slice-machine/src/legacy/lib/builders/SliceBuilder/VariationsList/VariationsList.module.css
@@ -1,0 +1,10 @@
+/* Hide scrollbar for Chrome, Safari and Opera */
+.hideScrollbar::-webkit-scrollbar {
+  display: none;
+}
+
+/* Hide scrollbar for IE, Edge and Firefox */
+.hideScrollbar {
+  -ms-overflow-style: none; /* IE and Edge */
+  scrollbar-width: none; /* Firefox */
+}
